separate dependencies & refactor
