+++ dependencies/__init__.py
@@ -1,0 +1,2 @@
+from .auth import *  # noqa
+from .get_repos import *  # noqa
