+++ libs/crypt.py
@@ -1,0 +1,7 @@
+from passlib.context import CryptContext
+
+Crypt = CryptContext(
+    schemes=["pbkdf2_sha256"],
+    default="pbkdf2_sha256",
+    pbkdf2_sha256__default_rounds=30000,
+)
