+++ repository/__init__.py
@@ -1,1 +1,5 @@
-from .dependencies import *  # noqa
+from .http import Http  # noqa
+from .metric_collector import MetricCollector  # noqa
+from .minio import Minio  # noqa
+from .postgres import Postgres  # noqa
+from .redis import Redis  # noqa
